+++ .github/github_workflow_scripts/check_protected_directories.py
