+++ .github/github_workflow_scripts/github_workflow_scripts_tests/check_if_partner_approved_label_test.py
@@ -4,11 +4,11 @@ import pytest
 @pytest.mark.parametrize(
     'support_levels, expected_support_level', [
         (
-            ['Utils/github_workflow_scripts/github_workflow_scripts_tests/test_files/Packs/Pack1/pack_metadata.json'],
+            ['.github/github_workflow_scripts/github_workflow_scripts_tests/test_files/Packs/Pack1/pack_metadata.json'],
             {'xsoar'}
         ),
         (
-            ['Utils/github_workflow_scripts/github_workflow_scripts_tests/test_files/Packs/Pack2/pack_metadata.json'],
+            ['.github/github_workflow_scripts/github_workflow_scripts_tests/test_files/Packs/Pack2/pack_metadata.json'],
             {'partner'}
         ),
     ]
@@ -24,5 +24,5 @@ def test_get_support_level(support_levels, expected_support_level):
         Then:
             - make sure the highest support level is always returned
     """
-    from Utils.github_workflow_scripts.check_if_partner_approved_label_exists import get_support_level
+    from github_workflow_scripts.check_if_partner_approved_label_exists import get_support_level
     assert get_support_level(support_levels) == expected_support_level
