+++ .github/workflows/check-contribution-form-filled.yml
@@ -29,5 +29,5 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: |
           echo "Checking if contribution form needs to be filled for PR: $PR_NUMBER"
-          cd Utils/github_workflow_scripts
+          cd .github/github_workflow_scripts
           poetry run ./check_if_needs_to_fill_contribution_form.py --pr_number $PR_NUMBER --github_token $GITHUB_TOKEN
