+++ .github/workflows/check-partner-approved-label.yml
@@ -33,5 +33,5 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: |
           echo "Checking label Partner-Approved for: $PR_NUMBER"
-          cd Utils/github_workflow_scripts
+          cd .github/github_workflow_scripts
           poetry run python check_if_partner_approved_label_exists.py --pr_number $PR_NUMBER --github_token $GITHUB_TOKEN
