+++ .github/workflows/protect-infra-directories.yml
@@ -25,4 +25,4 @@ jobs:
 
       - name: Check for changes in protected directories
         run: |
-              python Utils/github_workflow_scripts/check_protected_directories.py ${{ steps.changed-files.outputs.all_changed_files }}
+              python .github/github_workflow_scripts/check_protected_directories.py ${{ steps.changed-files.outputs.all_changed_files }}
