+++ .github/workflows/run-secrets-detection.yml
@@ -26,7 +26,7 @@ jobs:
           GOLD_SERVER_URL: ${{ secrets.GOLD_SERVER_URL_NG }}
         run: |
           echo "Run secrets detection for PR: $PR_NUMBER on branch: $BRANCH_NAME"
-          investigation_id=$(poetry run Utils/github_workflow_scripts/run_secrets_detection.py --pr_number $PR_NUMBER --branch_name $BRANCH_NAME --username $USERNAME --password $PASSWORD --gold_server_url $GOLD_SERVER_URL)
+          investigation_id=$(poetry run .github/github_workflow_scripts/run_secrets_detection.py --pr_number $PR_NUMBER --branch_name $BRANCH_NAME --username $USERNAME --password $PASSWORD --gold_server_url $GOLD_SERVER_URL)
           echo "INVESTIGATION_ID=$investigation_id" >> $GITHUB_ENV
       - name: Wait For Playbook To Finish
         env:
@@ -35,4 +35,4 @@ jobs:
           GOLD_SERVER_URL: ${{ secrets.GOLD_SERVER_URL_NG_RESULT }}
         run: |
           echo "Invastigation id is: $INVESTIGATION_ID "
-          poetry run python ./Utils/github_workflow_scripts/run_secrets_detection_get_playbook_status.py -i $INVESTIGATION_ID -k $GOLD_API_KEY -ai $AUTH_ID --gold_server_url $GOLD_SERVER_URL
+          poetry run python ./.github/github_workflow_scripts/run_secrets_detection_get_playbook_status.py -i $INVESTIGATION_ID -k $GOLD_API_KEY -ai $AUTH_ID --gold_server_url $GOLD_SERVER_URL
