+++ .github/workflows/sync-contribution-base-branch-on-change.yml
@@ -31,6 +31,6 @@ jobs:
           CONTENTBOT_GH_ADMIN_TOKEN: ${{ secrets.CONTENTBOT_GH_ADMIN_TOKEN }}
         run: |
           echo "Updating contribution base branch (contrib/*)"
-          cd Utils/github_workflow_scripts
+          cd .github/github_workflow_scripts
           poetry run ./sync_contrib_base.py --branch_name ${{ github.event.pull_request.base.ref }}
           echo "Finished updating base branch"
